+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Renamed the CSS class :current-year-selector to :year-selected.
 - Added support to shown the years that contain events in the years down-down menu (new CSS class :year-has-events).
 - Added support to search by Group and Url in the Search dialog.
+- Added a new checkbox called "Advanced" to the Search dialog, which wil hide the additional search options (saves screen space).
 - 
 
 ### Version 1.1.1: